04.Moduler/requirements.txt
requests==2.32.5
python-dotenv==1.0.1



#dette er en pom.xml / package.json
